tui: Extract shared save logic from AddScreen
build_part_from_form() and ValidationNotice now live in save.py so
the upcoming detail modal can reuse part construction and naming
without duplicating AddScreen internals.
